Dedupe Rails stub across specs into spec/support/rails_stub.rb
Eight specs each carried a near-identical `unless defined?(Rails); module
Rails; def self.env; 'test'; end; end; end` block to satisfy Apartment code
that reads Rails.env without booting the full framework. Extract to a single
shared helper auto-loaded from spec_helper.

`Rails.env` now returns an `ActiveSupport::StringInquirer` so `.test?` and
the string-comparison form both work — the integration spec for the
fixture-pool-lifecycle guard needed `.test?`. `Rails.root` (used by
abstract_adapter_spec for schema-file path resolution) is part of the same
helper.

`tenant_tagged_logging_spec.rb` retains its inline `stub_const` because it
needs per-example logger reassignment, not the static stub.
